+++ CHANGELOG.rst
@@ -26,8 +26,8 @@ Documentation
 -------------
 
 - Added a tutorial mirroring how ITN26 is written. (`#106 <https://github.com/LM-SAL/irispy/pull/106>`__)
-- Added an example of using `astroscrappy` to remove cosmic rays from IRIS data: :ref:`sphx_glr_generated_gallery_calibration_01_remove_spikes.py`. (`#106 <https://github.com/LM-SAL/irispy/pull/106>`__)
-- Added a how-to example showing how to remove dust from IRIS slit-jaw images with `irispy.sji.SJICube.remove_dust`: :ref:`sphx_glr_generated_gallery_calibration_03_remove_dust.py`. (`#111 <https://github.com/LM-SAL/irispy/pull/111>`__)
+- Added an example of using `astroscrappy` to remove cosmic rays from IRIS data: :ref:`sphx_glr_generated_gallery_calibration_01_remove_spikes_sg.py`. (`#106 <https://github.com/LM-SAL/irispy/pull/106>`__)
+- Added a how-to example showing how to remove dust from IRIS slit-jaw images with `irispy.sji.SJICube.remove_dust`: :ref:`sphx_glr_generated_gallery_calibration_04_remove_dust.py`. (`#111 <https://github.com/LM-SAL/irispy/pull/111>`__)
 - Added a double Gaussian Mg II k spectral fitting example: :ref:`sphx_glr_generated_gallery_analysis_07_mg_ii_two_gaussian_fitting.py`. (`#130 <https://github.com/LM-SAL/irispy/pull/130>`__)